Add files via upload
the python file I added is a testspace for working with .db files

this program allows you to upload specifically formatted txt files, and insert the stored data into a pre-existing database. The work is very rough, but the script functions as a basic proof-of-concept for a function which I plan to add to SYStool.py very soon.

the other files i attached are the test files I used to test this program with. notably, test.db and values.txt. if you look at test.db in  something like a db explorer, you will see that the values in values.txt are already in the database. this was done via the python script, demonstrating that it does indeed work as intended.

IMPORTANT TO NOTE: the python script assumes that you already have an ID system for your db. moreover, there is no format validation, so if your txt file does not match the specific format of your db, you will get errors out the wazzoo. this said, the script is surprisingly adaptable, working with all 2D databases, as far as I know.